+++ include/kernel.h
@@ -17,14 +17,14 @@ protected:
   real   R0;                                                    // Radius of root cell
   vect   Xperiodic;                                             // Coordinate offset of periodic image
 
-  std::vector<int>    keysHost;                                 // Offsets for rangeHost
-  std::vector<int>    rangeHost;                                // Offsets for sourceHost
-  std::vector<double> constHost;                                // Constants on host
-  std::vector<double> sourceHost;                               // Sources on host
-  std::vector<double> targetHost;                               // Targets on host
-  Map                 sourceBegin;                              // Define map for offset of source cells
-  Map                 sourceSize;                               // Define map for size of source cells
-  Map                 targetBegin;                              // Define map for offset of target cells
+  std::vector<int>   keysHost;                                  // Offsets for rangeHost
+  std::vector<int>   rangeHost;                                 // Offsets for sourceHost
+  std::vector<float> constHost;                                 // Constants on host
+  std::vector<float> sourceHost;                                // Sources on host
+  std::vector<float> targetHost;                                // Targets on host
+  Map                sourceBegin;                               // Define map for offset of source cells
+  Map                sourceSize;                                // Define map for size of source cells
+  Map                targetBegin;                               // Define map for offset of target cells
 
   int getLevel(bigint index) {                                  // Get level from cell index
     int level = -1;                                             // Initialize level counter
